+++ spec/ApiMailAdapter.spec.js
@@ -38,9 +38,12 @@ const config = {
             subjectPath: path.join(__dirname, 'templates/custom_email_subject.txt'),
             textPath: path.join(__dirname, 'templates/custom_email.txt'),
             htmlPath: path.join(__dirname, 'templates/custom_email.html'),
+            placeholders: {
+                appName: "TemplatePlaceholder"
+            },
             placeholderCallback: () => new Promise((resolve) => {
                 resolve({
-                    appName: 'ExampleApp'
+                    appName: 'CallbackPlaceholder'
                 });
             })
         },
@@ -49,7 +52,21 @@ const config = {
             textPath: path.join(__dirname, 'templates/custom_email.txt'),
             htmlPath: path.join(__dirname, 'templates/custom_email.html'),
             localeCallback: async () => { return 'de-AT'; }
-        }
+        },
+        customEmailWithPlaceholderCallbackAndLocaleCallback: {
+            subjectPath: path.join(__dirname, 'templates/custom_email_subject.txt'),
+            textPath: path.join(__dirname, 'templates/custom_email.txt'),
+            htmlPath: path.join(__dirname, 'templates/custom_email.html'),
+            placeholders: {
+                appName: "TemplatePlaceholder"
+            },
+            placeholderCallback: () => new Promise((resolve) => {
+                resolve({
+                    appName: 'CallbackPlaceholder'
+                });
+            }),
+            localeCallback: async () => { return 'de-AT'; }
+        },
     }
 };
 const examplePayload = {
@@ -400,7 +417,7 @@ describe('ApiMailAdapter', () => {
         });
     });
 
-    describe('validate placeholders', function () {
+    describe('placeholders', function () {
 
         it('returns valid placeholders', async () => {
             const adapter = new ApiMailAdapter(config);
@@ -412,6 +429,56 @@ describe('ApiMailAdapter', () => {
             const adapter = new ApiMailAdapter(config);
             const placeholders = 'invalid';
             expect(adapter._validatePlaceholders(placeholders)).toEqual({});
+        });
+
+        it('fills in the template placeholder without placeholder callback', async () => {
+            const adapter = new ApiMailAdapter(config);
+            const apiCallback = spyOn(adapter, 'apiCallback').and.callThrough();
+            const email = {
+                templateName: 'customEmailWithPlaceholderCallback',
+                recipient: 'to@example.com',
+                direct: true
+            }
+            const template = config.templates[email.templateName];
+            const templatePlaceholder = template.placeholders.appName;
+            const callbackPlaceholder = (await config.templates[email.templateName].placeholderCallback()).appName;
+            spyOn(template, 'placeholderCallback').and.callFake(() => {});
+            
+            await adapter._sendMail(email);
+            expect(apiCallback.calls.all()[0].args[0].payload.text).toContain(templatePlaceholder);
+            expect(apiCallback.calls.all()[0].args[0].payload.text).not.toContain(callbackPlaceholder);
+        });
+
+        it('overrides the template placeholder with the callback placeholder', async () => {
+            const adapter = new ApiMailAdapter(config);
+            const apiCallback = spyOn(adapter, 'apiCallback').and.callThrough();
+            const email = {
+                templateName: 'customEmailWithPlaceholderCallback',
+                recipient: 'to@example.com',
+                direct: true
+            }
+            const template = config.templates[email.templateName];
+            const templatePlaceholder = template.placeholders.appName;
+            const callbackPlaceholder = (await template.placeholderCallback()).appName;
+
+            await adapter._sendMail(email);
+            expect(apiCallback.calls.all()[0].args[0].payload.text).toContain(callbackPlaceholder);
+            expect(apiCallback.calls.all()[0].args[0].payload.text).not.toContain(templatePlaceholder);
+        });
+
+        it('makes user locale accessible in placeholder callback', async () => {
+            const adapter = new ApiMailAdapter(config);
+            const apiCallback = spyOn(adapter, 'apiCallback').and.callThrough();
+            const email = {
+                templateName: 'customEmailWithPlaceholderCallbackAndLocaleCallback',
+                recipient: 'to@example.com',
+                direct: true
+            }
+            const template = config.templates[email.templateName];
+            const locale = await template.localeCallback();
+            
+            await adapter._sendMail(email);
+            expect(apiCallback.calls.all()[0].args[0].locale).toBe(locale);
         });
     });